update: isolate ase-related calculator
