+++ src/py/mat3ra/made/tools/calculate/calculators.py
@@ -1,16 +1,13 @@
-from typing import Callable, List, Optional, Union
+from typing import Callable
 
 import numpy as np
 from mat3ra.made.material import Material
 from pydantic import BaseModel
 
 from ..analyze import get_surface_atom_indices
-from ..convert import from_ase
 from ..convert.utils import InterfacePartsEnum
 from ..enums import SurfaceTypes
 from ..modify import get_interface_part
-from ..third_party import ASEAtoms, ASECalculator, ASEFixAtoms, ASEFixedPlane, ase_all_changes
-from .constraints import RigidFilmXYConstraint
 from .interaction_functions import sum_of_inverse_distances_squared
 
 
@@ -123,112 +120,3 @@ class InterfaceMaterialCalculator(MaterialCalculator):
         substrate_coordinates_values = np.array(substrate_surface_atom_coordinates.values)
 
         return interaction_function(film_coordinates_values, substrate_coordinates_values)
-
-
-class FilmSubstrateDistanceASECalculator(ASECalculator):
-    """
-    ASE calculator that calculates the interaction energy between a film and substrate in an interface material.
-
-    Args:
-        shadowing_radius (float): The radius for atom to shadow underlying from being considered surface, in Angstroms.
-        force_constant (float): The force constant for the finite difference approximation of the forces.
-        is_substrate_fixed (bool): Whether to fix the substrate atoms.
-        is_z_axis_fixed (bool): Whether to fix atoms movement in the z direction.
-        symprec (float): The symmetry precision for the ASE calculator. This parameter determines the tolerance
-                         for symmetry operations, affecting the identification of equivalent atoms and the overall
-                         symmetry of the system. For more details, refer to the ASE documentation:
-                         https://wiki.fysik.dtu.dk/ase/ase/constraints.html#ase.constraints.FixSymmetry
-
-    Example usage:
-    ```python
-    from ase.optimize import BFGS
-    atoms = to_ase(material)
-    calc = SurfaceDistanceCalculator(shadowing_radius=2.5)
-
-    atoms.calc = calc
-    opt = BFGS(atoms)
-    opt.run(fmax=0.05)
-    ```
-    Args:
-        shadowing_radius (float): Radius for atoms shadowing underlying from being treated as a surface, in Angstroms.
-        force_constant (float): The force constant for the finite difference approximation of the
-    Note:
-        Built following https://wiki.fysik.dtu.dk/ase/development/calculators.html
-
-        The calculate method is responsible for computing the energy and forces (if requested).
-        Forces are estimated using a finite difference method, which is a simple approximation
-        and might not be the most accurate or efficient for all cases.
-    """
-
-    implemented_properties = ["energy", "forces"]
-
-    def __init__(
-        self,
-        shadowing_radius: float = 2.5,
-        is_substrate_fixed: bool = True,
-        is_z_axis_fixed: bool = True,
-        symprec: float = 0.01,
-        material_calculator: Union[MaterialCalculator, InterfaceMaterialCalculator] = InterfaceMaterialCalculator(),
-        **kwargs,
-    ):
-        super().__init__(**kwargs)
-        self.shadowing_radius = shadowing_radius
-        self.fix_substrate = is_substrate_fixed
-        self.fix_z = is_z_axis_fixed
-        self.symprec = symprec
-        self.material_calculator = material_calculator
-
-    def _add_constraints(self, atoms: ASEAtoms) -> ASEAtoms:
-        constraints: List[Union[ASEFixAtoms, ASEFixedPlane, RigidFilmXYConstraint]] = []
-        if self.fix_substrate:
-            substrate_indices = [i for i, tag in enumerate(atoms.get_tags()) if tag == 0]
-            constraints.append(ASEFixAtoms(indices=substrate_indices))
-        if self.fix_z:
-            all_indices = list(range(len(atoms)))
-            constraints.append(ASEFixedPlane(indices=all_indices, direction=[0, 0, 1]))
-
-        film_indices = [i for i, tag in enumerate(atoms.get_tags()) if tag == 1]
-        if film_indices:
-            constraints.append(RigidFilmXYConstraint(film_indices=film_indices))
-
-        atoms.set_constraint(constraints)
-        return atoms
-
-    def _calculate_forces(self, atoms: ASEAtoms, energy: float) -> np.ndarray:
-        forces = np.zeros((len(atoms), 3))
-        dx = 0.01
-        for i in range(len(atoms)):
-            for j in range(3):
-                atoms_plus = atoms.copy()
-                atoms_plus.positions[i, j] += dx
-                material_plus = Material(from_ase(atoms_plus))
-                energy_plus = self.material_calculator.get_energy(material_plus)
-
-                forces[i, j] = -(energy_plus - energy) / dx
-
-        return forces
-
-    def calculate(self, atoms: Optional[ASEAtoms] = None, properties=["energy"], system_changes=ase_all_changes):
-        if atoms is None:
-            atoms = self.atoms.copy()
-
-        atoms = self._add_constraints(atoms)
-        constraints = atoms.constraints
-
-        super().calculate(atoms, properties, system_changes)
-        material = Material(from_ase(atoms))
-        energy = self.material_calculator.get_energy(material)
-
-        self.results = {"energy": energy}
-
-        if "forces" in properties:
-            forces = self._calculate_forces(atoms, energy)
-            for constraint in constraints:
-                constraint.adjust_forces(atoms, forces)
-            self.results["forces"] = forces
-
-    def get_potential_energy(self, atoms=None, force_consistent=False):
-        return self.get_property("energy", atoms)
-
-    def get_forces(self, atoms=None):
-        return self.get_property("forces", atoms)
